+++ src/test/java/org/geowe/service/geometry/test/DivisionTest.java
@@ -15,9 +15,7 @@
  ******************************************************************************/
 package org.geowe.service.geometry.test;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.GenericType;
@@ -25,21 +23,18 @@ import javax.ws.rs.core.Response;
 import javax.ws.rs.core.Response.Status;
 
 import org.geowe.service.geometry.FlatGeometryBuilder;
-import org.geowe.service.geometry.JtsDivideService;
-import org.geowe.service.geometry.engine.GeometryExtractor;
-import org.geowe.service.geometry.engine.JTSGeoEngineerHelper;
+import org.geowe.service.model.DivisionData;
 import org.geowe.service.model.FlatGeometry;
-import org.geowe.service.model.OperationData;
 import org.geowe.service.model.error.ErrorEntity;
 import org.jboss.logging.Logger;
 import org.jboss.resteasy.client.jaxrs.ResteasyClient;
 import org.jboss.resteasy.client.jaxrs.ResteasyClientBuilder;
 import org.jboss.resteasy.client.jaxrs.ResteasyWebTarget;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 
-import com.vividsolutions.jts.geom.Geometry;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.vividsolutions.jts.util.Assert;
 
 public class DivisionTest {
@@ -49,140 +44,130 @@ public class DivisionTest {
 	private static final String SERVICE_URL = "http://127.0.0.1:8080/ggs/operations/jts/division";
 	private ResteasyClient restClient;
 	private ResteasyWebTarget target;
-	private static final String POLYGON_FC_WKT = "GEOMETRYCOLLECTION(POLYGON((-5.9685606135835565 38.333782907228816,-3.9141172545765723 38.31654526867596,-3.9360899108226595 37.52790590365409,-5.81475201986113 37.40582726607118,-5.9685606135835565 38.333782907228816)),POLYGON((-3.886651448642032 38.19360849133716,-3.5295957846435067 38.17201919361452,-3.8070005697500684 37.98392250848125,-3.886651448642032 38.19360849133716)))";
+	private static final String POLYGON_WKT = "POLYGON ((-5.9685606135835565 38.333782907228816,-3.9141172545765723 38.31654526867596,-3.9360899108226595 37.52790590365409,-5.81475201986113 37.40582726607118,-5.9685606135835565 38.333782907228816))";
 	private static final String LINE_WKT = "LINESTRING (-6.210259817917375 38.259416305016615,-5.188531302475292 37.679136788498276,-4.370049857309488 37.87018194277469,-3.463677787159337 38.272355099999935)";
-	private static final String LINE_2_WKT = "LINESTRING(-4.8663841287263905 37.78385163362393,-4.840634922188039 37.77910307500333,-4.7997795144805435 37.7928052283946)";
 	private static final String DIVISION_LINE_WKT = "LINESTRING(-4.85070171987867 37.840905948919264,-4.757317930832915 37.66935417624159)";
-	private static final String LINE_GC = "GEOMETRYCOLLECTION(LINESTRING(-4.8663841287263905 37.78385163362393,-4.840634922188039 37.77910307500333,-4.7997795144805435 37.7928052283946),LINESTRING(-6.210259817917375 38.2594163050166,-5.188531302475292 37.67913678849826,-4.370049857309488 37.870181942774686,-3.463677787159337 38.27235509999992),LINESTRING(-4.8592848030977125 37.76576530568103,-4.84898512048239 37.77268578071596,-4.843663617797742 37.76834359756804,-4.824437543582451 37.77634929940774,-4.818429395390217 37.771735949940556,-4.809331342413306 37.78069101241037,-4.802979871467189 37.781098034930324),LINESTRING(-4.853276654905389 37.75219386689201,-4.801434919074896 37.76332263045088,-4.801434919074896 37.76332263045088,-4.792851850228719 37.76277980277709),LINESTRING(-4.854306623166948 37.78584646538825,-4.814481183720922 37.79222244923771))";
-	private static final String CROSSED_LINES_1 ="LINESTRING(-4.8663841287263905 37.78385163362393,-4.840634922188039 37.77910307500333,-4.7997795144805435 37.7928052283946)";
-	private static final String CROSSED_LINES_2 ="LINESTRING(-4.854306623166948 37.78584646538825,-4.8308196263025325 37.78963116592093,-4.818460007164181 37.790512938540985,-4.812366028283342 37.7831871235283)";
+	private static final String GEOMETRY_COLLECTION ="GEOMETRYCOLLECTION(POLYGON((9.10125539332254 40.73595856620715,9.108894324595404 40.73914525646253,9.113529181772018 40.73452775766318,9.109838462168717 40.73017010572397,9.102199530894955 40.73049531348437,9.10125539332254 40.73595856620715)),LINESTRING(9.110267615610397 40.72782856293198,9.120052314095256 40.729454643056684,9.12236974268356 40.73426760734868,9.132326102545628 40.732316447569104,9.1341285470032 40.72847899974987),POINT(9.112928366953126 40.72054323623031))";
+	
 	@Before
 	public void setUp() throws Exception {
 		restClient = new ResteasyClientBuilder().build();
 		target = restClient.target(SERVICE_URL);
 	}
 
-	@Test
-	public void divideServiceTest() {
-		JtsDivideService service = new JtsDivideService();
-		Response response = service.dividePolygons(getOperationData());
-		@SuppressWarnings("unchecked")
-		List<FlatGeometry> dividedGeometries = (List<FlatGeometry>) response.getEntity();
-
-		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
-		Assert.isTrue(dividedGeometries.size() == 4);
-	}
-
-	@Test
-	public void dividePolygonsPostRequestTest() {
-		target = restClient.target(SERVICE_URL + "/polygons");
-		Response response = target.request().post(Entity.entity(getOperationData(), "application/json;charset=UTF-8"));
-
-		List<FlatGeometry> dividedGeometries = response.readEntity(new GenericType<List<FlatGeometry>>() {
-		});
-		response.close();
-		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
-		Assert.isTrue(dividedGeometries.size() == 4);
-
-	}
-
-	private OperationData getOperationData() {
-		OperationData opData = new OperationData();
-		HashSet<FlatGeometry> source = new HashSet<FlatGeometry>();
-		source.add(getFlatGeometry(POLYGON_FC_WKT));
-		opData.setSourceData(source);
-
-		HashSet<FlatGeometry> overlay = new HashSet<FlatGeometry>();
-		overlay.add(getFlatGeometry(LINE_WKT));
-		opData.setOverlayData(overlay);
-		return opData;
-	}
 
 	private FlatGeometry getFlatGeometry(String wkt) {
 		return new FlatGeometryBuilder().id("id-1").crs("WGS84").wkt(wkt).build();
 	}
-
-	@Test
-	public void dividePostRequestTopologyErrorTest() {
-		OperationData opData = new OperationData();
-
-		opData.setSourceData(new HashSet<FlatGeometry>(DataTestProvider.getThreeEntities()));
-		Set<FlatGeometry> overlayData = new HashSet<FlatGeometry>();
-		overlayData.add(DataTestProvider.createFlatGeometry(DataTestProvider.DIVISION_LINE_4_3POLYGONS));
-		opData.setOverlayData(overlayData);
-		target = restClient.target(SERVICE_URL + "/polygons");
-		Response response = target.request().post(Entity.entity(opData, "application/json;charset=UTF-8"));
-
-		ErrorEntity error = response.readEntity(ErrorEntity.class);
-		response.close();
-		log.info(error);
-		Assert.isTrue(response.getStatus() == Status.CONFLICT.getStatusCode());
-
-	}
-
-	@Test
-	public void dividePostRequestBadOverlayTest() {
-		OperationData opData = DataTestProvider.getPolygonsFCIntersectionData();
-		target = restClient.target(SERVICE_URL + "/polygons");
-		Response response = target.request().post(Entity.entity(opData, "application/json;charset=UTF-8"));
-
-		ErrorEntity error = response.readEntity(ErrorEntity.class);
-		response.close();
-		log.info(error);
-		Assert.isTrue(response.getStatus() == Status.BAD_REQUEST.getStatusCode());
-	}
-
-	@Test
-	public void divideLinesPostRequestTest() {
-		target = restClient.target(SERVICE_URL + "/lines");
-		Response response = target.request()
-				.post(Entity.entity(getLinesOperationData(LINE_GC), "application/json;charset=UTF-8"));
-
-		List<FlatGeometry> dividedGeometries = response
-				.readEntity(new GenericType<List<FlatGeometry>>() {});
-		response.close();
-		printGC(dividedGeometries);
-		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
-		Assert.isTrue(dividedGeometries.size() == 10);
-
-	}
 	
 	@Test
-	@Ignore
-	//TODO: Does not work with cross lines (not implemented)
-	public void divideCrossedLinesPostRequestTest() {
-		target = restClient.target(SERVICE_URL + "/lines");
-		OperationData opData = getLinesOperationData(CROSSED_LINES_1);
-		opData.getSourceData().add(getFlatGeometry(CROSSED_LINES_2));
+	public void divideLine(){
+		DivisionData divisionData = new DivisionData();
+		divisionData.setDivisionLine(getFlatGeometry(DIVISION_LINE_WKT));
+		divisionData.setGeomToDivide(getFlatGeometry(LINE_WKT));
+		printDivisionData(divisionData);
+		target = restClient.target(SERVICE_URL + "/line");
 		Response response = target.request()
-				.post(Entity.entity(opData, "application/json;charset=UTF-8"));
+				.post(Entity.entity(divisionData, "application/json;charset=UTF-8"));
 		
 		List<FlatGeometry> dividedGeometries = response
 				.readEntity(new GenericType<List<FlatGeometry>>() {});
 		response.close();
-		log.info(dividedGeometries.size());
-		printGC(dividedGeometries);
+		
 		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
-		Assert.isTrue(dividedGeometries.size() == 4);
-
+		Assert.isTrue(dividedGeometries.size() == 2);
+	}
+	
+	@Test
+	public void divideLineBadDivisionLine(){
+		DivisionData divisionData = new DivisionData();
+		divisionData.setDivisionLine(getFlatGeometry(POLYGON_WKT));
+		divisionData.setGeomToDivide(getFlatGeometry(LINE_WKT));
+		
+		target = restClient.target(SERVICE_URL + "/line");
+		Response response = target.request()
+				.post(Entity.entity(divisionData, "application/json;charset=UTF-8"));
+		
+		ErrorEntity error= response
+				.readEntity(ErrorEntity.class);
+		response.close();
+		Assert.isTrue(response.getStatus() == Status.BAD_REQUEST.getStatusCode());
+	}
+	
+	@Test
+	public void divideLineBadLineToDivide(){
+		DivisionData divisionData = new DivisionData();
+		divisionData.setDivisionLine(getFlatGeometry(DIVISION_LINE_WKT));
+		divisionData.setGeomToDivide(getFlatGeometry(POLYGON_WKT));
+		
+		target = restClient.target(SERVICE_URL + "/line");
+		Response response = target.request()
+				.post(Entity.entity(divisionData, "application/json;charset=UTF-8"));
+		
+		ErrorEntity error= response
+				.readEntity(ErrorEntity.class);
+		response.close();
+		Assert.isTrue(response.getStatus() == Status.BAD_REQUEST.getStatusCode());
+	}
+	
+	@Test
+	public void dividePolygon(){
+		DivisionData divisionData = new DivisionData();
+		divisionData.setDivisionLine(getFlatGeometry(LINE_WKT));
+		divisionData.setGeomToDivide(getFlatGeometry(POLYGON_WKT));
+		printDivisionData(divisionData);
+		target = restClient.target(SERVICE_URL + "/polygon");
+		Response response = target.request()
+				.post(Entity.entity(divisionData, "application/json;charset=UTF-8"));
+		
+		List<FlatGeometry> dividedGeometries = response
+				.readEntity(new GenericType<List<FlatGeometry>>() {});
+		response.close();
+		
+		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
+		Assert.isTrue(dividedGeometries.size() == 2);
+	}
+	
+	@Test
+	public void divideBadPolygonToDivide(){
+		DivisionData divisionData = new DivisionData();
+		divisionData.setDivisionLine(getFlatGeometry(LINE_WKT));
+		divisionData.setGeomToDivide(getFlatGeometry(DIVISION_LINE_WKT));
+		
+		target = restClient.target(SERVICE_URL + "/polygon");
+		Response response = target.request()
+				.post(Entity.entity(divisionData, "application/json;charset=UTF-8"));
+		
+		ErrorEntity error= response
+				.readEntity(ErrorEntity.class);
+		response.close();
+		Assert.isTrue(response.getStatus() == Status.BAD_REQUEST.getStatusCode());
+	}
+	
+	
+	@Test
+	public void decomposeFeatureCollection(){
+				
+		Response response = target.request()
+				.post(Entity.entity(getFlatGeometry(GEOMETRY_COLLECTION), "application/json;charset=UTF-8"));
+		
+		List<FlatGeometry> dividedGeometries = response
+				.readEntity(new GenericType<List<FlatGeometry>>() {});
+		response.close();
+		Assert.isTrue(response.getStatus() == Status.CREATED.getStatusCode());
+		Assert.isTrue(dividedGeometries.size() == 3);
+	}
+	
+	private void printDivisionData(DivisionData divisionData){
+		ObjectMapper mapper = new ObjectMapper();
+		try {
+			log.info(mapper.writeValueAsString(divisionData));
+		} catch (JsonProcessingException e) {
+			// TODO Auto-generated catch block
+			e.printStackTrace();
+		}
 	}
 
-	private void printGC(List<FlatGeometry> fgeoms) {
-		Geometry geom = new GeometryExtractor()
-				.createGeometryCollection(new JTSGeoEngineerHelper().toGeometries(fgeoms));
-		log.info("GEOMETRY COLLECTION: " + geom);
-	}
-
-	private OperationData getLinesOperationData(String sourceDataWkt) {
-		OperationData opData = new OperationData();
-		HashSet<FlatGeometry> source = new HashSet<FlatGeometry>();
-		source.add(getFlatGeometry(sourceDataWkt));
-		opData.setSourceData(source);
-
-		HashSet<FlatGeometry> overlay = new HashSet<FlatGeometry>();
-		overlay.add(getFlatGeometry(DIVISION_LINE_WKT));
-		opData.setOverlayData(overlay);
-		return opData;
-	}
+	
 
 }